Move shoe routes to new file.
