+++ api/fitness/app/app.py
@@ -13,6 +13,7 @@ from fitness.models import Run
 from .constants import DEFAULT_START, DEFAULT_END
 from .dependencies import all_runs, update_new_runs_only
 from .metrics import router as metrics_router
+from .shoe_routes import router as shoe_router
 from fitness.utils.timezone import convert_runs_to_user_timezone
 
 RunSortBy = Literal[
@@ -23,6 +24,7 @@ SortOrder = Literal["asc", "desc"]
 
 app = FastAPI()
 app.include_router(metrics_router)
+app.include_router(shoe_router)
 
 app.add_middleware(
     CORSMiddleware,
